+++ ElvUI_LuckyoneUI/Core/Core.lua
@@ -48,13 +48,6 @@ function L1UI:Setup_Details()
 	L1UI:Print('Details profile has been set.')
 end
 
--- Load Gnosis Profile
-function L1UI:Setup_Gnosis()
-	if not IsAddOnLoaded('Gnosis') then return end
-	L1UI:Get_Gnosis_Profile()
-	L1UI:Print('Gnosis profile has been set.')
-end
-
 -- Load OmniCD Profile
 function L1UI:Setup_OmniCD()
 	if not IsAddOnLoaded('OmniCD') then return end
@@ -143,10 +136,10 @@ function L1UI:SetupPrivate()
 	E.private["skins"]["parchmentRemoverEnable"] = true
 
 	if L1UI.Retail then
-		E.private["install_complete"] = "12.37"
+		E.private["install_complete"] = "12.44"
 		E.private["general"]["totemBar"] = false
 	elseif L1UI.TBC then
-		E.private["install_complete"] = "2.10"
+		E.private["install_complete"] = "2.17"
 		E.private["general"]["totemBar"] = true
 	elseif L1UI.Classic then
 		E.private["install_complete"] = "1.43"